+++ tests/corpus/go/go-select-001/input.go
@@ -1,0 +1,12 @@
+package main
+
+func recv(a, b chan int) int {
+	select {
+	case x := <-a:
+		return x
+	case y := <-b:
+		return y
+	default:
+		return 0
+	}
+}